+++ app/src/main/java/kr/ac/hansung/demap/ui/createfolder/List_onClick_interface.kt
@@ -1,4 +1,4 @@
-package kr.ac.hansung.demap
+package kr.ac.hansung.demap.ui.createfolder
 
 //어댑터에서 데이터 넘겨받기 위한 인터페이스
 interface List_onClick_interface {
